アンケートのページの移動とホームの作成
