+++ src/pages/questionnaire/index.tsx
@@ -1,0 +1,67 @@
+import Head from "next/head";
+// import { Inter } from "next/font/google";
+import ComboList from "@/components/selectList/comboList";
+import { useState } from "react";
+import { ComboItem } from "@/interfaces";
+
+// const inter = Inter({ subsets: ["latin"] });
+
+export default function Home() {
+  const initComboItem: ComboItem = { id: 0, name: "" };
+  const [selectedItem, setSelectedItem] = useState<ComboItem>(initComboItem);
+  const [selected2Item, setSelected2Item] = useState<ComboItem>(initComboItem);
+
+  const data: ComboItem[] = [
+    { id: 1, name: "あああ" },
+    { id: 2, name: "いいい" },
+    { id: 3, name: "ううう" },
+    { id: 4, name: "えええ" },
+    { id: 5, name: "その他" },
+  ];
+
+  const data2: ComboItem[] = [
+    { id: 1, name: "はい" },
+    { id: 2, name: "いいえ" },
+  ];
+
+  // useEffect(() => {
+  //   console.log(selectedItem);
+  // }, [selectedItem]);
+
+  return (
+    <>
+      <Head>
+        <title>Next Practice</title>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <h1>あああ</h1>
+      {/* <p>
+        {selectedItem === initComboItem
+          ? "何も選択されていません"
+          : `選択されたものは: ${selectedItem.id}:${selectedItem.name}`}
+      </p> */}
+      <ComboList
+        name={"質問1"}
+        items={data}
+        length={data.length}
+        setItem={(item) => setSelectedItem(item)}
+        stVisible={selectedItem === initComboItem}
+      />
+      {/* <p>
+        {selected2Item === initComboItem
+          ? "何も選択されていません"
+          : `選択されたものは: ${selected2Item.id}:${selected2Item.name}`}
+      </p> */}
+      <ComboList
+        name={"質問2"}
+        items={data2}
+        length={data2.length}
+        setItem={(item) => setSelected2Item(item)}
+        stVisible={selected2Item === initComboItem}
+      />
+    </>
+  );
+}
